Changes Atlas ObjectID type to unsigned int, fixes 64-bit build errors caused by size_t.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@11186 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/AtlasUI/ScenarioEditor/Tools/Common/MiscState.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2009 Wildfire Games.
+/* Copyright (C) 2012 Wildfire Games.
  * This file is part of 0 A.D.
  *
  * 0 A.D. is free software: you can redistribute it and/or modify
@@ -22,7 +22,7 @@
 
 namespace AtlasMessage
 {
-	typedef size_t ObjectID;
+	typedef unsigned int ObjectID;
 }
 
 extern wxString g_SelectedTexture;
